Switch from detailed view only on double click on border
Adjusted the logic to switch from the detailed view to only listen to
double clicks on the border. This was done to avoid the user
accidentally switching the view when trying to interact with the view.

Moves the check whether a position is inside the node border to a
separate method to avoid code duplication.